Adjusting gencode
