+++ gencode/docs/event_discovery.html
@@ -1341,6 +1341,8 @@
     <a href="#points_pattern1_ancillary" onclick="anchorLink('points_pattern1_ancillary')">ancillary</a></div><span class="badge badge-dark value-type">Type: object</span><br/>
 <span class="description"><p>Arbitrary blob of json associated with this point</p>
 </span>
+
+    
             
 
             
@@ -1381,8 +1383,10 @@
             />
         </svg>
     <a href="#ancillary" onclick="anchorLink('ancillary')">ancillary</a></div><span class="badge badge-dark value-type">Type: object</span><br/>
-<span class="description"><p>Arbitrary blob of json associated with this device</p>
+<span class="description"><p>Arbitrary blob of json associated with this point</p>
 </span>
+
+
