--- gencode/java/udmi/schema/Ancillary.java
@@ -1,39 +1,0 @@
-
-package udmi.schema;
-
-import javax.annotation.processing.Generated;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-
-
-/**
- * Arbitrary blob of json associated with this device
- * 
- */
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({
-
-})
-@Generated("jsonschema2pojo")
-public class Ancillary {
-
-
-    @Override
-    public int hashCode() {
-        int result = 1;
-        return result;
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        if (other == this) {
-            return true;
-        }
-        if ((other instanceof Ancillary) == false) {
-            return false;
-        }
-        Ancillary rhs = ((Ancillary) other);
-        return true;
-    }
-
-}
