+++ schema/ancillary_properties.json
@@ -1,0 +1,6 @@
+{
+  "title": "Ancillary Properties",
+  "description": "Arbitrary blob of json associated with this point",
+  "existingJavaType": "java.util.HashMap<String, Object>",
+  "type": "object"
+}
