+++ schema/event_discovery.json
@@ -50,8 +50,7 @@
       }
     },
     "ancillary": {
-      "description": "Arbitrary blob of json associated with this device",
-      "type": "object"
+      "$ref": "file:ancillary_properties.json"
     },
     "hardware": {
       "$ref": "file:state_system_hardware.json"
